@@ -4,7 +4,7 @@
 
   <groupId>com.jedk1</groupId>
   <artifactId>jedcore</artifactId>
-  <version>2.13.2-PK1.11.1</version>
+  <version>2.15.0-PK1.11.1</version>
   <packaging>jar</packaging>
   <name>JedCore</name>
 
@@ -13,10 +13,6 @@
       <id>spigot-repo</id>
       <url>https://hub.spigotmc.org/nexus/content/repositories/snapshots/</url>
     </repository>
-      <repository>
-        <id>jitpack.io</id>
-        <url>https://jitpack.io</url>
-      </repository>
   </repositories>
 
   <dependencies>
@@ -24,15 +20,14 @@
     <dependency>
       <groupId>org.spigotmc</groupId>
       <artifactId>spigot-api</artifactId>
-      <version>1.16.5-R0.1-SNAPSHOT</version>
+      <version>1.20.2-R0.1-SNAPSHOT</version>
       <scope>provided</scope>
     </dependency>
     <!-- ProjectKorra -->
     <dependency>
-      <groupId>com.github.ProjectKorra</groupId>
-      <artifactId>ProjectKorra</artifactId>
-      <version>v1.11.1</version>
-      <scope>provided</scope>
+      <groupId>com.projectkorra</groupId>
+      <artifactId>projectkorra</artifactId>
+      <version>1.11.1</version>
     </dependency>
   </dependencies>
 
@@ -56,8 +51,8 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>3.8.0</version>
         <configuration>
-          <source>1.8</source>
-          <target>1.8</target>
+          <source>9</source>
+          <target>9</target>
         </configuration>
       </plugin>
       <plugin>
