@@ -12,7 +12,6 @@ import com.projectkorra.projectkorra.ability.util.Collision;
 import com.projectkorra.projectkorra.attribute.Attribute;
 import com.projectkorra.projectkorra.region.RegionProtection;
 import com.projectkorra.projectkorra.util.DamageHandler;
-
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.configuration.ConfigurationSection;
@@ -29,25 +28,26 @@ import java.util.stream.Collectors;
 
 public class DaggerThrow extends ChiAbility implements AddonAbility {
 	private static final List<AbilityInteraction> INTERACTIONS = new ArrayList<>();
-	private static boolean particles;
-	private static double damage;
 
+	private final List<Arrow> arrows = new ArrayList<>();
+
+	private boolean limitEnabled;
+	private boolean requireArrows;
+	private boolean allowPickup;
+	private boolean particles;
 	private long endTime;
 	private int shots = 1;
+	private int hits = 0;
+
+	@Attribute(Attribute.DAMAGE)
+	private double damage;
 	@Attribute(Attribute.COOLDOWN)
 	private long cooldown;
-	private boolean limitEnabled;
 	@Attribute("MaxShots")
 	private int maxShots;
-	private int hits = 0;
-	private final List<Arrow> arrows = new ArrayList<>();
 
 	public DaggerThrow(Player player) {
 		super(player);
-
-		if (this instanceof DamageAbility) {
-			return;
-		}
 
 		if (!bPlayer.canBend(this)) {
 			return;
@@ -66,6 +66,7 @@ public class DaggerThrow extends ChiAbility implements AddonAbility {
 		setFields();
 
 		start();
+
 		if (!isRemoved()) {
 			shootArrow();
 		}
@@ -79,6 +80,8 @@ public class DaggerThrow extends ChiAbility implements AddonAbility {
 		maxShots = config.getInt("Abilities.Chi.DaggerThrow.MaxDaggers.Amount");
 		particles = config.getBoolean("Abilities.Chi.DaggerThrow.ParticleTrail");
 		damage = config.getDouble("Abilities.Chi.DaggerThrow.Damage");
+		requireArrows = config.getBoolean("Abilities.Chi.DaggerThrow.RequireArrows");
+		allowPickup = config.getBoolean("Abilities.Chi.DaggerThrow.AllowPickup");
 
 		loadInteractions();
 	}
@@ -90,6 +93,7 @@ public class DaggerThrow extends ChiAbility implements AddonAbility {
 
 		ConfigurationSection config = JedCoreConfig.getConfig(this.player);
 		ConfigurationSection section = config.getConfigurationSection(path);
+
 		for (String abilityName : section.getKeys(false)) {
 			INTERACTIONS.add(new AbilityInteraction(abilityName));
 		}
@@ -101,11 +105,13 @@ public class DaggerThrow extends ChiAbility implements AddonAbility {
 			remove();
 			return;
 		}
+
 		if (System.currentTimeMillis() > endTime) {
 			bPlayer.addCooldown(this);
 			remove();
 			return;
 		}
+
 		if (shots > maxShots && limitEnabled) {
 			bPlayer.addCooldown(this);
 			remove();
@@ -113,74 +119,79 @@ public class DaggerThrow extends ChiAbility implements AddonAbility {
 	}
 
 	private void shootArrow() {
-		if (JCMethods.removeItemFromInventory(player, Material.ARROW, 1)) {
-			shots++;
-			Location location = player.getEyeLocation();
+		shots++;
+		Location location = player.getEyeLocation();
 
-			Vector vector = location.toVector().
-					add(location.getDirection().multiply(2.5)).
-					toLocation(location.getWorld()).toVector().
-					subtract(player.getEyeLocation().toVector());
+		Vector vector = location.toVector().
+				add(location.getDirection().multiply(2.5)).
+				toLocation(location.getWorld()).toVector().
+				subtract(player.getEyeLocation().toVector());
 
-			Arrow arrow = player.launchProjectile(Arrow.class);
-			arrow.setVelocity(vector);
-			arrow.getLocation().setDirection(vector);
-			arrow.setKnockbackStrength(0);
-			arrow.setBounce(false);
-			arrow.setMetadata("daggerthrow", new FixedMetadataValue(JedCore.plugin, "1"));
+		if (requireArrows) JCMethods.removeItemFromInventory(player, Material.ARROW, 1);
 
-			if (particles) {
-				arrow.setCritical(true);
-			}
+		Arrow arrow = player.launchProjectile(Arrow.class);
+		arrow.setVelocity(vector);
+		arrow.getLocation().setDirection(vector);
+		arrow.setKnockbackStrength(0);
+		arrow.setBounce(false);
+		arrow.setMetadata("daggerthrow", new FixedMetadataValue(JedCore.plugin, "1"));
 
-			arrows.add(arrow);
-			endTime = System.currentTimeMillis() + 500;
-			bPlayer.addCooldown("DaggerThrowShot", 100);
+		if (!allowPickup) arrow.setPickupStatus(Arrow.PickupStatus.DISALLOWED);
+
+		if (particles) {
+			arrow.setCritical(true);
 		}
+
+		arrows.add(arrow);
+		endTime = System.currentTimeMillis() + 500;
+		bPlayer.addCooldown("DaggerThrowShot", 100);
 	}
 
-	public static void damageEntityFromArrow(LivingEntity entity, Arrow arrow) {
-		if (RegionProtection.isRegionProtected((Player) arrow.getShooter(), arrow.getLocation(), "DaggerThrow")) {
-			return;
-		}
+	public void damageEntityFromArrow(LivingEntity entity, Arrow arrow) {
+		if (!(arrow.getShooter() instanceof Player shooter)) return;
+
+		if (RegionProtection.isRegionProtected(shooter, arrow.getLocation(), "DaggerThrow")) return;
 
 		arrow.setVelocity(new Vector(0, 0, 0));
 		entity.setNoDamageTicks(0);
+
 		double prevHealth = entity.getHealth();
-		Player shooter = (Player) arrow.getShooter();
-		DamageAbility da = new DamageAbility(shooter);
-		DamageHandler.damageEntity(entity, damage, da);
-		da.remove();
+
+		DamageHandler.damageEntity(entity, damage, this);
+
 		if (prevHealth > entity.getHealth()) {
 			arrow.remove();
 		}
 
-		if (!(entity instanceof Player)) {
+		if (!(entity instanceof Player target)) {
 			return;
 		}
 
-		DaggerThrow dt = CoreAbility.getAbility(shooter, DaggerThrow.class);
-		if (dt == null) {
+		DaggerThrow daggerThrow = CoreAbility.getAbility(shooter, DaggerThrow.class);
+		if (daggerThrow == null) {
 			return;
 		}
 
-		++dt.hits;
-
-		Player target = (Player)entity;
-		BendingPlayer bPlayer = BendingPlayer.getBendingPlayer(target);
+		daggerThrow.hits++;
+		BendingPlayer targetBPlayer = BendingPlayer.getBendingPlayer(target);
 
 		for (AbilityInteraction interaction : INTERACTIONS) {
-			if (!interaction.enabled) continue;
-			if (dt.hits < interaction.hitRequirement) continue;
+			if (!interaction.enabled || daggerThrow.hits < interaction.hitRequirement) {
+				continue;
+			}
 
 			CoreAbility abilityDefinition = AbilitySelector.getAbility(interaction.name);
-			if (abilityDefinition == null) continue;
+			if (abilityDefinition == null) {
+				continue;
+			}
 
 			CoreAbility ability = CoreAbility.getAbility(target, abilityDefinition.getClass());
-			if (ability == null) continue;
+			if (ability == null) {
+				continue;
+			}
 
 			ability.remove();
-			bPlayer.addCooldown(ability, interaction.cooldown);
+			targetBPlayer.addCooldown(ability, interaction.cooldown);
 		}
 	}
 	
@@ -250,11 +261,11 @@ public class DaggerThrow extends ChiAbility implements AddonAbility {
 		return "* JedCore Addon *\n" + config.getString("Abilities.Chi.DaggerThrow.Description");
 	}
 
-	public static boolean hasParticleTrail() {
+	public boolean hasParticleTrail() {
 		return particles;
 	}
 
-	public static double getDamage() {
+	public double getDamage() {
 		return damage;
 	}
 
@@ -316,44 +327,6 @@ public class DaggerThrow extends ChiAbility implements AddonAbility {
 	public boolean isEnabled() {
 		ConfigurationSection config = JedCoreConfig.getConfig(this.player);
 		return config.getBoolean("Abilities.Chi.DaggerThrow.Enabled");
-	}
-	
-	public static class DamageAbility extends DaggerThrow {
-		
-		public DamageAbility(Player player) {
-			super(player);
-			start();
-		}
-
-		@Override
-		public long getCooldown() {
-			return 0;
-		}
-
-		@Override
-		public Location getLocation() {
-			return null;
-		}
-
-		@Override
-		public String getName() {
-			return "DaggerThrow";
-		}
-
-		@Override
-		public boolean isHarmlessAbility() {
-			return false;
-		}
-
-		@Override
-		public boolean isSneakAbility() {
-			return false;
-		}
-
-		@Override
-		public void progress() {
-			remove();
-		}
 	}
 
 	private class AbilityInteraction {
