@@ -4,21 +4,19 @@ import java.util.*;
 
 import com.jedk1.jedcore.util.*;
 import com.projectkorra.projectkorra.ability.ElementalAbility;
+import com.projectkorra.projectkorra.ability.WaterAbility;
 import com.projectkorra.projectkorra.region.RegionProtection;
-import org.bukkit.Effect;
-import org.bukkit.Location;
-import org.bukkit.Material;
-import org.bukkit.World;
+import org.bukkit.*;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.block.data.Levelled;
+import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.InventoryHolder;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.util.Vector;
 
 import com.jedk1.jedcore.configuration.JedCoreConfig;
-import com.jedk1.jedcore.scoreboard.BendingBoard;
 import com.projectkorra.projectkorra.GeneralMethods;
 import com.projectkorra.projectkorra.ProjectKorra;
 import com.projectkorra.projectkorra.ability.CoreAbility;
@@ -26,6 +24,8 @@ import com.projectkorra.projectkorra.ability.util.ComboManager;
 import com.projectkorra.projectkorra.util.TempBlock;
 
 public class JCMethods {
+
+	// todo: either use PKs isPlant or the registry/tags (with config)
 	private static final ArrayList<Material> SMALL_PLANTS = new ArrayList<Material>(){{
 		addAll(Arrays.asList(Material.GRASS, Material.FERN, Material.POPPY, Material.DANDELION, Material.OAK_SAPLING,
 				Material.SPRUCE_SAPLING, Material.BIRCH_SAPLING, Material.JUNGLE_SAPLING, Material.ACACIA_SAPLING,
@@ -35,6 +35,7 @@ public class JCMethods {
 				Material.BROWN_MUSHROOM, Material.PUMPKIN_STEM, Material.MELON_STEM, Material.WHEAT, Material.TALL_GRASS,
 				Material.BEETROOTS, Material.CARROTS, Material.POTATOES, Material.CRIMSON_FUNGUS, Material.WARPED_FUNGUS,
 				Material.BAMBOO, Material.BAMBOO_SAPLING));
+
 		int serverVersion = GeneralMethods.getMCVersion();
 		if (serverVersion >= 1170) {
 			add(Material.getMaterial("AZALEA"));
@@ -308,6 +309,7 @@ public class JCMethods {
 		return false;
 	}
 
+	// todo: see SMALL_PLANTS
 	public static boolean isDoublePlant(Material material) {
 		return material == Material.SUNFLOWER || material == Material.LILAC || material == Material.TALL_GRASS ||
 				material == Material.LARGE_FERN || material == Material.ROSE_BUSH || material == Material.PEONY;
@@ -318,23 +320,37 @@ public class JCMethods {
 	}
 
 	public static boolean isSmallPlant(Material material) {
-		return SMALL_PLANTS.contains(material);
+		return WaterAbility.isPlant(material);
+		//return SMALL_PLANTS.contains(material);
+	}
+
+	public static void displayColoredParticles(String hex, Location location, int amount, double offsetX, double offsetY, double offsetZ, double extra) {
+		displayColoredParticles(hex, location, amount, offsetX, offsetY, offsetZ, extra, 255);
+	}
+
+	public static void displayColoredParticles(String hex, Location location, int amount, double offsetX, double offsetY, double offsetZ, double extra, int alpha) {
+		JedCore.plugin.getParticleAdapter().displayColoredParticles(hex, location, amount, offsetX, offsetY, offsetZ, extra, alpha);
+	}
+
+	public static void emitLight(Location loc) {
+		ConfigurationSection config = JedCoreConfig.getConfig((Player)null);
+		if (config.getBoolean("Properties.Fire.DynamicLight.Enabled")) {
+			int brightness = config.getInt("Properties.Fire.DynamicLight.Brightness");
+			long keepAlive = config.getLong("Properties.Fire.DynamicLight.KeepAlive");
+
+			LightManager.createLight(loc).brightness(brightness).timeUntilFadeout(keepAlive).emit();
+		}
 	}
 
 	public static void reload() {
 		JedCore.log.info("JedCore Reloaded.");
 		JedCore.plugin.reloadConfig();
 		JedCore.logDebug = JedCoreConfig.getConfig((World)null).getBoolean("Properties.LogDebug");
-		JedCoreConfig.board.reloadConfig();
 		CoreAbility.registerPluginAbilities(JedCore.plugin, "com.jedk1.jedcore.ability");
 		registerDisabledWorlds();
 		registerCombos();
 		RegenTempBlock.revertAll();
-		BendingBoard.setFields();
-		BendingBoard.updateOnline();
 		JedCore.plugin.initializeCollisions();
 		FireTick.loadMethod();
-
-		BendingBoard.loadOtherCooldowns();
 	}
 }
