@@ -8,12 +8,14 @@ import com.jedk1.jedcore.configuration.JedCoreConfig;
 import com.jedk1.jedcore.util.BlockUtil;
 import com.projectkorra.projectkorra.ability.AddonAbility;
 import com.projectkorra.projectkorra.ability.EarthAbility;
-import com.projectkorra.projectkorra.ability.ElementalAbility;
 import com.projectkorra.projectkorra.ability.util.Collision;
 import com.projectkorra.projectkorra.attribute.Attribute;
+import com.projectkorra.projectkorra.earthbending.passive.DensityShift;
 import com.projectkorra.projectkorra.region.RegionProtection;
-import com.projectkorra.projectkorra.util.*;
-
+import com.projectkorra.projectkorra.util.DamageHandler;
+import com.projectkorra.projectkorra.util.ParticleEffect;
+import com.projectkorra.projectkorra.util.TempBlock;
+import com.projectkorra.projectkorra.util.TempFallingBlock;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
@@ -24,17 +26,27 @@ import org.bukkit.entity.Player;
 import org.bukkit.util.Vector;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
-import java.util.Random;
+import java.util.Set;
+import java.util.UUID;
+import java.util.concurrent.ThreadLocalRandom;
 
 import static java.util.stream.Collectors.toList;
 
 public class EarthKick extends EarthAbility implements AddonAbility {
 	private final List<TempFallingBlock> temps = new ArrayList<>();
+	private final Set<UUID> hitEntities = new HashSet<>();
 
 	private BlockData materialData;
 	private Location location;
-	private final Random rand = new Random();
+	private Block block;
+	private boolean multipleHits;
+	private int sourceRange;
+	private int spread;
+	private double velocity;
+	private boolean allowMetal;
+	private boolean replaceSource;
 
 	@Attribute(Attribute.COOLDOWN)
 	private long cooldown;
@@ -42,9 +54,10 @@ public class EarthKick extends EarthAbility implements AddonAbility {
 	private int earthBlocks;
 	@Attribute(Attribute.DAMAGE)
 	private double damage;
+	@Attribute(Attribute.DAMAGE)
+	private double metalDmg;
 	@Attribute("CollisionRadius")
 	private double entityCollisionRadius;
-	private Block block;
 
 	public EarthKick(Player player) {
 		super(player);
@@ -54,7 +67,9 @@ public class EarthKick extends EarthAbility implements AddonAbility {
 		}
 
 		setFields();
+
 		location = player.getLocation();
+
 		if ((player.getLocation().getPitch() > -5) && prepare()) {
 			if (RegionProtection.isRegionProtected(this, block.getLocation())) {
 				return;
@@ -69,8 +84,15 @@ public class EarthKick extends EarthAbility implements AddonAbility {
 		
 		cooldown = config.getLong("Abilities.Earth.EarthKick.Cooldown");
 		earthBlocks = config.getInt("Abilities.Earth.EarthKick.EarthBlocks");
-		damage = config.getDouble("Abilities.Earth.EarthKick.Damage");
+		damage = config.getDouble("Abilities.Earth.EarthKick.Damage.Normal");
+		metalDmg = config.getDouble("Abilities.Earth.EarthKick.Damage.Metal");
 		entityCollisionRadius = config.getDouble("Abilities.Earth.EarthKick.EntityCollisionRadius");
+		multipleHits = config.getBoolean("Abilities.Earth.EarthKick.MultipleHits");
+		sourceRange = config.getInt("Abilities.Earth.EarthKick.SourceRange");
+		spread = config.getInt("Abilities.Earth.EarthKick.Spread");
+		velocity = config.getDouble("Abilities.Earth.EarthKick.Velocity");
+		allowMetal = config.getBoolean("Abilities.Earth.EarthKick.AllowMetal");
+		replaceSource = config.getBoolean("Abilities.Earth.EarthKick.ReplaceSource");
 
 		if (entityCollisionRadius < 1.0) {
 			entityCollisionRadius = 1.0;
@@ -78,12 +100,25 @@ public class EarthKick extends EarthAbility implements AddonAbility {
 	}
 
 	private boolean prepare() {
-		block = player.getTargetBlock(getTransparentMaterialSet(), 2);
-		if (!isEarthbendable(player, block)){
+		block = player.getTargetBlock(getTransparentMaterialSet(), sourceRange);
+
+		if (EarthAbility.getMovedEarth().containsKey(block)) {
 			return false;
 		}
 
-		if (block != null && !isMetal(block)) {
+		if (!isEarthbendable(player, block)) {
+			return false;
+		}
+
+		if (TempBlock.isTempBlock(block)) {
+			TempBlock.get(block).revertBlock();
+		}
+
+		if (DensityShift.isPassiveSand(block)) {
+			DensityShift.revertSand(block);
+		}
+
+		if (block != null && (allowMetal || !isMetal(block))) {
 			materialData = block.getBlockData().clone();
 			location.setX(block.getX() + 0.5);
 			location.setY(block.getY());
@@ -108,6 +143,7 @@ public class EarthKick extends EarthAbility implements AddonAbility {
 		}
 
 		bPlayer.addCooldown(this);
+
 		track();
 
 		if (temps.isEmpty()) {
@@ -116,12 +152,15 @@ public class EarthKick extends EarthAbility implements AddonAbility {
 	}
 
 	private void launchBlocks() {
-		if (getMovedEarth().containsKey(block)) {
-			block.setType(Material.AIR);
-		}
-		if (block.getType() != Material.AIR) {
-			TempBlock air = new TempBlock(block, Material.AIR);
-			air.setRevertTime(5000L);
+		if (replaceSource) {
+			if (getMovedEarth().containsKey(block)) {
+				block.setType(Material.AIR);
+			}
+
+			if (block.getType() != Material.AIR) {
+				TempBlock air = new TempBlock(block, Material.AIR);
+				air.setRevertTime(5000L);
+			}
 		}
 
 		location.setPitch(0);
@@ -137,13 +176,15 @@ public class EarthKick extends EarthAbility implements AddonAbility {
 
 		playEarthbendingSound(location);
 
+		ThreadLocalRandom rand = ThreadLocalRandom.current();
+
 		for (int i = 0; i < earthBlocks; i++) {
-			location.setYaw(yaw + (rand.nextInt((20 - -20) + 1) + -20));
+			location.setYaw(yaw + rand.nextInt((spread * 2) + 1) - spread);
 			location.setPitch(rand.nextInt(25) - 45);
 
 			Vector v = location.clone().add(0, 0.8, 0).getDirection().normalize();
 			Location location1 = location.clone().add(new Vector(v.getX() * 2, v.getY(), v.getZ() * 2));
-			Vector dir = location1.setDirection(location.getDirection()).getDirection();
+			Vector dir = location1.setDirection(location.getDirection()).getDirection().multiply(velocity);
 
 			temps.add(new TempFallingBlock(location, materialData, dir, this));
 		}
@@ -168,7 +209,10 @@ public class EarthKick extends EarthAbility implements AddonAbility {
 			AABB collider = BlockUtil.getFallingBlockBoundsFull(fb).scale(entityCollisionRadius * 2.0);
 
 			CollisionDetector.checkEntityCollisions(player, collider, (entity) -> {
-				DamageHandler.damageEntity(entity, damage, this);
+				UUID uuid = entity.getUniqueId();
+				if (this.multipleHits || hitEntities.add(uuid)) {
+					DamageHandler.damageEntity(entity, isMetal(fb.getBlockData().getMaterial()) ? metalDmg : damage, this);
+				}
 				return false;
 			});
 		}
@@ -263,6 +307,14 @@ public class EarthKick extends EarthAbility implements AddonAbility {
 
 	public void setDamage(double damage) {
 		this.damage = damage;
+	}
+
+	public double getMetalDmg() {
+		return metalDmg;
+	}
+
+	public void setMetalDmg(double metalDmg) {
+		this.metalDmg = metalDmg;
 	}
 
 	public double getEntityCollisionRadius() {
