@@ -8,7 +8,6 @@ import com.projectkorra.projectkorra.ability.EarthAbility;
 import com.projectkorra.projectkorra.earthbending.EarthArmor;
 import com.projectkorra.projectkorra.util.TempArmor;
 import com.projectkorra.projectkorra.util.TempPotionEffect;
-
 import org.bukkit.Color;
 import org.bukkit.Location;
 import org.bukkit.Material;
@@ -17,7 +16,6 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.LeatherArmorMeta;
 import org.bukkit.potion.PotionEffect;
-import org.bukkit.potion.PotionEffectType;
 
 import java.util.Arrays;
 import java.util.List;
@@ -97,7 +95,8 @@ public class MetalArmor extends EarthAbility implements AddonAbility {
 				}
 
 				player.getInventory().setArmorContents(armors);
-				PotionEffect resistance = new PotionEffect(PotionEffectType.DAMAGE_RESISTANCE, resistDuration / 50, resistStrength - 1);
+
+				PotionEffect resistance = JedCore.plugin.getPotionEffectAdapter().getResistanceEffect(resistDuration, resistStrength);
 				new TempPotionEffect(player, resistance);
 			}
 
