@@ -4,6 +4,7 @@ import com.jedk1.jedcore.JedCore;
 import com.jedk1.jedcore.configuration.JedCoreConfig;
 import com.projectkorra.projectkorra.GeneralMethods;
 import com.projectkorra.projectkorra.ability.AddonAbility;
+import com.projectkorra.projectkorra.ability.EarthAbility;
 import com.projectkorra.projectkorra.ability.MetalAbility;
 import com.projectkorra.projectkorra.attribute.Attribute;
 import com.projectkorra.projectkorra.region.RegionProtection;
@@ -13,7 +14,6 @@ import com.projectkorra.projectkorra.util.DamageHandler;
 import com.projectkorra.projectkorra.util.ParticleEffect;
 import com.projectkorra.projectkorra.util.TempBlock;
 import com.projectkorra.projectkorra.util.TempFallingBlock;
-
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.Sound;
@@ -28,7 +28,12 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.util.Vector;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.ListIterator;
+import java.util.Random;
 
 public class MetalFragments extends MetalAbility implements AddonAbility {
 
@@ -42,6 +47,7 @@ public class MetalFragments extends MetalAbility implements AddonAbility {
 	private double damage;
 	@Attribute(Attribute.COOLDOWN)
 	private long cooldown;
+	private double velocity;
 
 	public List<Block> sources = new ArrayList<>();
 	private final List<Item> thrownFragments = new ArrayList<>();
@@ -88,6 +94,7 @@ public class MetalFragments extends MetalAbility implements AddonAbility {
 		maxFragments = config.getInt("Abilities.Earth.MetalFragments.MaxFragments");
 		damage = config.getDouble("Abilities.Earth.MetalFragments.Damage");
 		cooldown = config.getInt("Abilities.Earth.MetalFragments.Cooldown");
+		velocity = config.getDouble("Abilities.Earth.MetalFragments.Velocity");
 	}
 
 	public static void shootFragment(Player player) {
@@ -130,7 +137,7 @@ public class MetalFragments extends MetalAbility implements AddonAbility {
 
 		Item ii = player.getWorld().dropItemNaturally(source.getLocation().getBlock().getRelative(GeneralMethods.getCardinalDirection(direction)).getLocation(), is);
 		ii.setPickupDelay(Integer.MAX_VALUE);
-		ii.setVelocity(direction.multiply(2).normalize());
+		ii.setVelocity(direction.normalize().multiply(velocity));
 		playMetalbendingSound(ii.getLocation());
 		thrownFragments.add(ii);
 
@@ -179,11 +186,16 @@ public class MetalFragments extends MetalAbility implements AddonAbility {
 		if (block == null)
 			return false;
 
+		if (EarthAbility.getMovedEarth().containsKey(block))
+			return false;
+
 		return isMetal(block);
 	}
 
 	public Block selectSource() {
 		Block block = BlockSource.getEarthSourceBlock(player, selectRange, ClickType.SHIFT_DOWN);
+		if (EarthAbility.getMovedEarth().containsKey(block))
+			return null;
 		if (isMetal(block))
 			return block;
 		return null;
