@@ -13,7 +13,6 @@ import com.projectkorra.projectkorra.attribute.Attribute;
 import com.projectkorra.projectkorra.region.RegionProtection;
 import com.projectkorra.projectkorra.util.DamageHandler;
 import com.projectkorra.projectkorra.util.ParticleEffect;
-
 import com.projectkorra.projectkorra.util.TempFallingBlock;
 import org.bukkit.Location;
 import org.bukkit.Material;
@@ -220,6 +219,8 @@ public class FireComet extends FireAbility implements AddonAbility {
 		playFirebendingParticles(location, 20, Math.random(), Math.random(), Math.random());
 		ParticleEffect.FIREWORKS_SPARK.display(location, 20,  Math.random(), Math.random(), Math.random(), 0.5);
 
+		JCMethods.emitLight(location);
+
 		location.getWorld().playSound(location, (rand.nextBoolean()) ? Sound.ENTITY_FIREWORK_ROCKET_BLAST : Sound.ENTITY_FIREWORK_ROCKET_BLAST_FAR, 5F, 1F);
 		location.getWorld().playSound(location, Sound.ENTITY_GENERIC_EXPLODE, 5F, 0.8F);
 
@@ -267,6 +268,9 @@ public class FireComet extends FireAbility implements AddonAbility {
 		ParticleEffect.SMOKE_LARGE.display(location.clone().add(v), 1, 0, 0, 0, 0.02);
 		ParticleEffect.SMOKE_LARGE.display(location.clone().add(v1), 1, 0, 0, 0, 0.02);
 
+		JCMethods.emitLight(location.clone().add(v));
+		JCMethods.emitLight(location.clone().add(v1));
+
 		if (this.angle == 360) {
 			this.angle = 0;
 		}
@@ -283,6 +287,7 @@ public class FireComet extends FireAbility implements AddonAbility {
 //					ParticleEffect.FLAME.display(l, 1, 0, 0, 0, 0.02);
 //				}
 				playFirebendingParticles(l, 1, 0, 0, 0);
+				JCMethods.emitLight(l);
 			}
 		}
 
@@ -300,6 +305,7 @@ public class FireComet extends FireAbility implements AddonAbility {
 //					ParticleEffect.FLAME.display(l, 1, 0, 0, 0, 0.05);
 //				}
 				playFirebendingParticles(l, 1, 0, 0, 0);
+				JCMethods.emitLight(l);
 			}
 		}
 
@@ -330,6 +336,9 @@ public class FireComet extends FireAbility implements AddonAbility {
 			playFirebendingParticles(location.clone().add(v1), 1, 0, 0, 0);
 			ParticleEffect.SMOKE_LARGE.display(location.clone().add(v), 1, 0, 0, 0, 0.02);
 			ParticleEffect.SMOKE_LARGE.display(location.clone().add(v1), 1, 0, 0, 0, 0.02);
+
+			JCMethods.emitLight(location.clone().add(v));
+			JCMethods.emitLight(location.clone().add(v1));
 		}
 
 		if (point == 360) {
