@@ -1,26 +1,39 @@
 package com.jedk1.jedcore.ability.waterbending;
 
-import java.util.*;
-
-import com.jedk1.jedcore.configuration.JedCoreConfig;
-import com.projectkorra.projectkorra.ability.ElementalAbility;
-import com.projectkorra.projectkorra.attribute.Attribute;
-import com.projectkorra.projectkorra.region.RegionProtection;
-import org.bukkit.Location;
-import org.bukkit.configuration.ConfigurationSection;
-import org.bukkit.entity.*;
-import org.bukkit.potion.Potion;
-import org.bukkit.potion.PotionType;
-import org.bukkit.util.Vector;
-
 import com.jedk1.jedcore.JedCore;
+import com.jedk1.jedcore.configuration.JedCoreConfig;
 import com.projectkorra.projectkorra.BendingPlayer;
 import com.projectkorra.projectkorra.GeneralMethods;
 import com.projectkorra.projectkorra.ability.AddonAbility;
 import com.projectkorra.projectkorra.ability.AirAbility;
 import com.projectkorra.projectkorra.ability.BloodAbility;
+import com.projectkorra.projectkorra.ability.ElementalAbility;
+import com.projectkorra.projectkorra.attribute.Attribute;
 import com.projectkorra.projectkorra.command.Commands;
+import com.projectkorra.projectkorra.region.RegionProtection;
 import com.projectkorra.projectkorra.util.DamageHandler;
+import org.bukkit.Location;
+import org.bukkit.Material;
+import org.bukkit.configuration.ConfigurationSection;
+import org.bukkit.entity.Arrow;
+import org.bukkit.entity.Blaze;
+import org.bukkit.entity.Creature;
+import org.bukkit.entity.Creeper;
+import org.bukkit.entity.Entity;
+import org.bukkit.entity.Fireball;
+import org.bukkit.entity.Ghast;
+import org.bukkit.entity.LivingEntity;
+import org.bukkit.entity.Player;
+import org.bukkit.entity.Skeleton;
+import org.bukkit.entity.ThrownPotion;
+import org.bukkit.entity.Witch;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.PotionMeta;
+import org.bukkit.util.Vector;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
 
 public class BloodPuppet extends BloodAbility implements AddonAbility {
 
@@ -264,9 +277,11 @@ public class BloodPuppet extends BloodAbility implements AddonAbility {
 				if (target instanceof LivingEntity) {
 					LivingEntity e = (LivingEntity) target;
 					ThrownPotion tp = missmagus.launchProjectile(ThrownPotion.class, GeneralMethods.getDirection(puppet.getEyeLocation(), e.getEyeLocation()));
-					Potion potion = new Potion(PotionType.INSTANT_DAMAGE);
-					potion.setSplash(true);
-					tp.setItem(potion.toItemStack(1));
+					ItemStack potionItem = new ItemStack(Material.SPLASH_POTION, 1);
+					PotionMeta potion = (PotionMeta) potionItem.getItemMeta();
+					potion.setBasePotionType(JedCore.plugin.getPotionEffectAdapter().getHarmingPotionType());
+					potionItem.setItemMeta(potion);
+					tp.setItem(potionItem);
 					tp.setVelocity(GeneralMethods.getDirection(puppet.getEyeLocation(), e.getEyeLocation()).multiply(0.125));
 					tp.setShooter(puppet);
 					if (e instanceof Creature)
