@@ -1,6 +1,7 @@
 package com.jedk1.jedcore.ability.waterbending;
 
 import com.jedk1.jedcore.JedCore;
+import com.jedk1.jedcore.ability.firebending.FirePunch;
 import com.jedk1.jedcore.configuration.JedCoreConfig;
 import com.projectkorra.projectkorra.GeneralMethods;
 import com.projectkorra.projectkorra.ability.AddonAbility;
@@ -15,25 +16,35 @@ import org.bukkit.entity.ArmorStand;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Player;
-import org.bukkit.potion.PotionEffect;
+import org.bukkit.inventory.MainHand;
 import org.bukkit.potion.PotionEffectType;
 
 public class IceClaws extends IceAbility implements AddonAbility {
 
 	@Attribute(Attribute.COOLDOWN)
-	private long cooldown;
+	private long punchCooldown;
+	@Attribute(Attribute.COOLDOWN)
+	private long throwCooldown;
 	@Attribute(Attribute.CHARGE_DURATION)
 	private long chargeUp;
-	private int slowDur;
+	private int punchSlowDur;
+	private int throwSlowDur;
+	private int punchSlownessLevel;
+	private int throwSlownessLevel;
 	@Attribute(Attribute.DAMAGE)
-	private double damage;
+	private double punchDamage;
+	@Attribute(Attribute.DAMAGE)
+	private double throwDamage;
 	@Attribute(Attribute.RANGE)
 	private double range;
 	private boolean throwable;
+	private double throwSpeed;
 
 	private Location head;
 	private Location origin;
 	private boolean launched;
+
+	private Boolean iceInMainHand = null;
 
 	public IceClaws(Player player) {
 		super(player);
@@ -56,19 +67,29 @@ public class IceClaws extends IceAbility implements AddonAbility {
 	public void setFields() {
 		ConfigurationSection config = JedCoreConfig.getConfig(this.player);
 
-		cooldown = config.getLong("Abilities.Water.IceClaws.Cooldown");
-		chargeUp = config.getLong("Abilities.Water.IceClaws.ChargeTime");
-		slowDur = config.getInt("Abilities.Water.IceClaws.SlowDuration")/50;
-		damage = config.getDouble("Abilities.Water.IceClaws.Damage");
-		range = config.getDouble("Abilities.Water.IceClaws.Range");
-		throwable = config.getBoolean("Abilities.Water.IceClaws.Throwable");
-		
+		punchCooldown = config.getLong("Abilities.Water.IceClaws.Punch.Cooldown", 4000);
+		punchDamage = config.getDouble("Abilities.Water.IceClaws.Punch.Damage", 2.0);
+		punchSlownessLevel = config.getInt("Abilities.Water.IceClaws.Punch.Slowness", 3);
+		punchSlowDur = config.getInt("Abilities.Water.IceClaws.Punch.SlowDuration", 5000);
+
+		throwCooldown = config.getLong("Abilities.Water.IceClaws.Throw.Cooldown", 4000);
+		throwDamage = config.getDouble("Abilities.Water.IceClaws.Throw.Damage", 2.0);
+		throwSlownessLevel = config.getInt("Abilities.Water.IceClaws.Throw.Slowness", 3);
+		throwSlowDur = config.getInt("Abilities.Water.IceClaws.Throw.SlowDuration", 5000);
+		throwSpeed = config.getDouble("Abilities.Water.IceClaws.Throw.Speed", 1.0);
+
+		chargeUp = config.getLong("Abilities.Water.IceClaws.ChargeTime", 1000);
+		range = config.getDouble("Abilities.Water.IceClaws.Range", 10);
+		throwable = config.getBoolean("Abilities.Water.IceClaws.Throwable", true);
+
 		applyModifiers();
 	}
 	
 	private void applyModifiers() {
-		cooldown -= ((long) getNightFactor(cooldown) - cooldown);
-		damage = getNightFactor(damage);
+		punchCooldown -= ((long) getNightFactor(punchCooldown) - punchCooldown);
+		throwCooldown -= ((long) getNightFactor(throwCooldown) - throwCooldown);
+		punchDamage = getNightFactor(punchDamage);
+		throwDamage = getNightFactor(throwDamage);
 		range = getNightFactor(range);
 	}
 
@@ -97,9 +118,27 @@ public class IceClaws extends IceAbility implements AddonAbility {
 		}
 	}
 
+
+	public static void swapHands(Player player) {
+		ConfigurationSection config = JedCoreConfig.getConfig(player);
+		if (!config.getBoolean("Abilities.Water.IceClaws.AllowHandSwap", true)) return;
+		IceClaws ic = getAbility(player, IceClaws.class);
+		if (ic == null)
+			return;
+		if (ic.iceInMainHand == null)
+			ic.iceInMainHand = true;
+		else ic.iceInMainHand = !ic.iceInMainHand;
+	}
+
+	public Location getRightHandPos() {
+		return (player.getMainHand() == MainHand.RIGHT == ((iceInMainHand == null) || iceInMainHand) ?
+				GeneralMethods.getRightSide(player.getLocation(), .55) :
+				GeneralMethods.getLeftSide(player.getLocation(), .55)).add(0, 1.2, 0);
+	}
+
 	public boolean shoot() {
-		for (double i = 0; i < 1; i+=.5) {
-			head.add(origin.clone().getDirection().multiply(.5));
+		for (double i = 0; i < 1; i += throwSpeed) {
+			head.add(origin.clone().getDirection().multiply(throwSpeed));
 			if (origin.distance(head) >= range) return false;
 			if (!isTransparent(head.getBlock())) return false;
 			GeneralMethods.displayColoredParticle("66FFFF", head);
@@ -107,13 +146,12 @@ public class IceClaws extends IceAbility implements AddonAbility {
 			ParticleEffect.SNOW_SHOVEL.display(head, 1, 0, 0, 0, 0);
 			for (Entity entity : GeneralMethods.getEntitiesAroundPoint(head, 1.5)) {
 				if (entity instanceof LivingEntity && entity.getEntityId() != player.getEntityId() && !(entity instanceof ArmorStand)) {
-					freezeEntity((LivingEntity) entity);
+					freezeEntity((LivingEntity) entity, false);
 					return false;
 				}
 			}
 		}
 		return true;
-
 	}
 
 	public static void throwClaws(Player player) {
@@ -125,10 +163,6 @@ public class IceClaws extends IceAbility implements AddonAbility {
 				ic.head = ic.origin.clone();
 			}
 		}
-	}
-
-	public Location getRightHandPos() {
-		return GeneralMethods.getRightSide(player.getLocation(), .55).add(0, 1.2, 0);
 	}
 
 	private void displayClaws() {
@@ -151,18 +185,24 @@ public class IceClaws extends IceAbility implements AddonAbility {
 	}
 
 	private void freezeEntity(LivingEntity entity) {
+		freezeEntity(entity, true);
+	}
+
+	private void freezeEntity(LivingEntity entity, boolean isPunch) {
 		if (entity.hasPotionEffect(PotionEffectType.SPEED)) {
 			entity.removePotionEffect(PotionEffectType.SPEED);
 		}
-		entity.addPotionEffect(new PotionEffect(PotionEffectType.SLOW, slowDur, 3));
+		int duration = isPunch ? punchSlowDur : throwSlowDur;
+		int level = isPunch ? punchSlownessLevel : throwSlownessLevel;
+		entity.addPotionEffect(JedCore.plugin.getPotionEffectAdapter().getSlownessEffect(duration, level));
 		bPlayer.addCooldown(this);
 		remove();
-		DamageHandler.damageEntity(entity, damage, this);
+		DamageHandler.damageEntity(entity, isPunch ? punchDamage : throwDamage, this);
 	}
 	
 	@Override
 	public long getCooldown() {
-		return cooldown;
+		return launched ? throwCooldown : punchCooldown;
 	}
 
 	@Override
@@ -201,33 +241,18 @@ public class IceClaws extends IceAbility implements AddonAbility {
 		return "* JedCore Addon *\n" + config.getString("Abilities.Water.IceClaws.Description");
 	}
 
-	public void setCooldown(long cooldown) {
-		this.cooldown = cooldown;
-	}
-
-	public long getChargeUp() {
-		return chargeUp;
-	}
-
-	public void setChargeUp(long chargeUp) {
-		this.chargeUp = chargeUp;
-	}
-
-	public int getSlowDuration() {
-		return slowDur;
-	}
-
-	public void setSlowDuration(int slowDuration) {
-		this.slowDur = slowDuration;
-	}
-
-	public double getDamage() {
-		return damage;
-	}
-
-	public void setDamage(double damage) {
-		this.damage = damage;
-	}
+	public void setPunchCooldown(long punchCooldown) { this.punchCooldown = punchCooldown; }
+	public long getPunchCooldown() { return punchCooldown; }
+	public void setThrowCooldown(long throwCooldown) { this.throwCooldown = throwCooldown; }
+	public long getThrowCooldown() { return throwCooldown; }
+	public void setPunchDamage(double punchDamage) { this.punchDamage = punchDamage; }
+	public double getPunchDamage() { return punchDamage; }
+	public void setThrowDamage(double throwDamage) { this.throwDamage = throwDamage; }
+	public double getThrowDamage() { return throwDamage; }
+	public void setPunchSlowDur(int punchSlowDur) { this.punchSlowDur = punchSlowDur; }
+	public int getPunchSlowDur() { return punchSlowDur; }
+	public void setThrowSlowDur(int throwSlowDur) { this.throwSlowDur = throwSlowDur; }
+	public int getThrowSlowDur() { return throwSlowDur; }
 
 	public double getRange() {
 		return range;
@@ -279,5 +304,29 @@ public class IceClaws extends IceAbility implements AddonAbility {
 	public boolean isEnabled() {
 		ConfigurationSection config = JedCoreConfig.getConfig(this.player);
 		return config.getBoolean("Abilities.Water.IceClaws.Enabled");
+	}
+
+	public int getPunchSlownessLevel() {
+		return punchSlownessLevel;
+	}
+
+	public void setPunchSlownessLevel(int punchSlownessLevel) {
+		this.punchSlownessLevel = punchSlownessLevel;
+	}
+
+	public int getThrowSlownessLevel() {
+		return throwSlownessLevel;
+	}
+
+	public void setThrowSlownessLevel(int throwSlownessLevel) {
+		this.throwSlownessLevel = throwSlownessLevel;
+	}
+
+	public double getThrowSpeed() {
+		return throwSpeed;
+	}
+
+	public void setThrowSpeed(double throwSpeed) {
+		this.throwSpeed = throwSpeed;
 	}
 }
