@@ -4,7 +4,6 @@ import com.jedk1.jedcore.JedCore;
 import com.projectkorra.projectkorra.GeneralMethods;
 import com.projectkorra.projectkorra.ability.Ability;
 import com.projectkorra.projectkorra.util.DamageHandler;
-
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Player;
@@ -15,12 +14,12 @@ import java.util.concurrent.ConcurrentHashMap;
 
 public class ThrownEntityTracker {
 
-	public static ConcurrentHashMap<Entity, ThrownEntityTracker> instances = new ConcurrentHashMap<Entity, ThrownEntityTracker>();
+	public static ConcurrentHashMap<Entity, ThrownEntityTracker> instances = new ConcurrentHashMap<>();
 	public static boolean collisions = JedCore.plugin.getConfig().getBoolean("Properties.MobCollisions.Enabled");
 	private long delay;
-	private long fireTime;
-	private Entity entity;
-	private Player instigator;
+	private final long fireTime;
+	private final Entity entity;
+	private final Player instigator;
 	private Vector thisVelocity;
 	private Ability ability;
 
@@ -38,18 +37,18 @@ public class ThrownEntityTracker {
 		if (System.currentTimeMillis() < fireTime + delay) {
 			return;
 		}
+
 		if (!collisions || entity.isOnGround()) {
 			remove();
 			return;
 		}
+
 		thisVelocity = entity.getVelocity().clone();
+
 		List<Entity> nearby = GeneralMethods.getEntitiesAroundPoint(entity.getLocation(), 2D);
-		if (nearby.contains(entity)) {
-			nearby.remove(entity);
-		}
-		if (nearby.contains(instigator)) {
-			nearby.remove(instigator);
-		}
+        nearby.remove(entity);
+        nearby.remove(instigator);
+
 		if (nearby.size() != 0) {
 			entity.setVelocity(thisVelocity.multiply(0.5D));
 			for(Entity e : nearby){
